+++ projects/_03_independent_cqc/_07_estimate_filled_posts_by_job_role/fargate/validate_01_merge.py
@@ -7,6 +7,9 @@ import polars as pl
 from polars_utils import utils
 from polars_utils.validation import actions as vl
 from polars_utils.validation.constants import GLOBAL_ACTIONS, GLOBAL_THRESHOLDS
+from projects._03_independent_cqc._07_estimate_filled_posts_by_job_role.fargate.utils.merge_utils import (
+    reduced_data_filter_expr,
+)
 from projects._03_independent_cqc._07_estimate_filled_posts_by_job_role.fargate.utils.utils import (
     CategoricalColumnTypes,
 )
@@ -32,6 +35,7 @@ VALIDATION_COLS_TO_IMPORT = [
 
 IND_CQC_ESTIMATES_COLS_TO_IMPORT = [
     IndCqcColumns.location_id,
+    IndCqcColumns.cqc_location_import_date,
 ]
 
 CQC_EARLIEST_IMPORT_DATE = date(2013, 3, 1)
@@ -78,7 +82,7 @@ def main(
     compare_df = utils.read_parquet(
         source=f"s3://{bucket_name}/{compare_path}",
         selected_columns=IND_CQC_ESTIMATES_COLS_TO_IMPORT,
-    )
+    ).filter(reduced_data_filter_expr())
     expected_row_count = compare_df.height * len(
         AscwdsWorkerValueLabelsJobGroup.all_roles()
     )